desc：repair some bug

+++ app/src/main/java/com/jiwenjie/cocomusic/ui/MainActivity.kt
@@ -4,7 +4,9 @@ import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.support.v7.widget.LinearLayoutManager
+import android.util.Log
 import android.view.KeyEvent
+import com.jiwenjie.basepart.utils.LogUtils
 import com.jiwenjie.basepart.utils.ToastUtils
 import com.jiwenjie.cocomusic.R
 import com.jiwenjie.cocomusic.aidl.Music
@@ -12,6 +14,7 @@ import com.jiwenjie.cocomusic.common.Constants
 import com.jiwenjie.cocomusic.playservice.PlayManager
 import com.jiwenjie.cocomusic.ui.adapter.MusicListAdapter
 import com.jiwenjie.cocomusic.utils.SongLoader
+import com.squareup.leakcanary.AnalyzedHeap.save
 import kotlinx.android.synthetic.main.activity_main.*
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
@@ -22,6 +25,8 @@ class MainActivity : PlayBaseActivity() {
 
     private val beanList by lazy { ArrayList<Music>() }
     private val adapter by lazy { MusicListAdapter(this, beanList) }
+
+    private var currentMusic: Music? = null
 
     companion object {
        fun runActivity(activity: Activity) {
@@ -36,11 +41,20 @@ class MainActivity : PlayBaseActivity() {
        recyclerView.adapter = adapter
        adapter.setOnItemClickListener { position, view ->
            if (beanList.size == 0) return@setOnItemClickListener
-           val id = Random().nextInt(beanList.size)
-           PlayManager.play(id, beanList, Constants.PLAYLIST_LOCAL_ID)
-//           adapter.notifyDataSetChanged()
+           // 处理用户重复点击一首歌曲的时候每次都重新开始播放
+           if (currentMusic != null) {
+               val clickMusic = beanList[position]
+               if (currentMusic != clickMusic) {
+                   currentMusic = clickMusic
+                   PlayManager.play(position, beanList, Constants.PLAYLIST_LOCAL_ID)
+               }
+           } else {
+               currentMusic = beanList[position]
+               PlayManager.play(position, beanList, Constants.PLAYLIST_LOCAL_ID)
+           }
        }
 
+       // 异步读取本地歌曲
        doAsync {
            val data = SongLoader.getAllLocalSongs(this@MainActivity)
            uiThread {
@@ -54,18 +68,23 @@ class MainActivity : PlayBaseActivity() {
    }
 
     // 双击退出程序
-    var prePressTime = 0L
+    var prePressTime = 0.toLong()
 
     override fun onKeyDown(keyCode: Int, event: KeyEvent?): Boolean {
-        if (System.currentTimeMillis() - prePressTime > 2000) {
-            prePressTime = System.currentTimeMillis()
-            ToastUtils.showToast(this, "再按一次退出程序")
+        if (keyCode == KeyEvent.KEYCODE_BACK) {
+            if (System.currentTimeMillis() - prePressTime > 2000) {
+                prePressTime = System.currentTimeMillis()
+                ToastUtils.showToast(this, "再按一次退出程序")
+            } else {
+                finish()
+                android.os.Process.killProcess(android.os.Process.myUid())
+                System.exit(0)
+            }
+            return false
         } else {
-            finish()
-            android.os.Process.killProcess(android.os.Process.myUid())
-            System.exit(0)
+            // 点击音量键加减的时候也会响应该方法，所以在这里处理，防止点击音量键会导致应用退出
+            return super.onKeyDown(keyCode, event)
         }
-        return false
     }
 
    override fun getLayoutId(): Int = R.layout.activity_main
